+++ .gitignore
@@ -1,2 +1,2 @@
 __pycache__
-# */*__pycache__
+*/*__pycache__
